requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
line-bot-sdk==3.13.0
anthropic==0.40.0
notion-client==2.2.1
python-dotenv==1.0.1
httpx==0.27.2
